backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
python-docx
google-genai
python-dotenv
